Move combination class to correct package
Revert ReductiveTreeWalker
