+++ generator/src/main/java/com/scottlogic/deg/generator/generation/combination/CombinationProducer.java
@@ -1,7 +1,5 @@
 package com.scottlogic.deg.generator.generation.combination;
 
-import com.scottlogic.deg.generator.decisiontree.rule_strategy.Combination;
-
 import java.util.stream.Stream;
 
 public interface CombinationProducer {
